LPFG-929: move org dir into csrs
