+++ src/applicationContext.ts
@@ -49,7 +49,7 @@ import {DateRangeCommand} from './controllers/command/dateRangeCommand'
 import {DateRangeCommandFactory} from './controllers/command/factory/dateRangeCommandFactory'
 import {DateRange} from './learning-catalogue/model/dateRange'
 import {DateRangeFactory} from './learning-catalogue/model/factory/dateRangeFactory'
-import {OrganisationController} from './csrs/organisationController'
+import {OrganisationController} from './controllers/organisationController'
 import {Csrs} from './csrs'
 import {OrganisationalUnitFactory} from './csrs/model/organisationalUnitFactory'
 import {LearnerRecord} from './learner-record'
