+++ test/unit/controllers/organisationControllerTest.ts
@@ -5,7 +5,7 @@ import {expect} from 'chai'
 import * as sinonChai from 'sinon-chai'
 import {Request, Response} from 'express'
 import * as sinon from 'sinon'
-import {OrganisationController} from '../../../src/csrs/organisationController'
+import {OrganisationController} from '../../../src/controllers/organisationController'
 import {Csrs} from '../../../src/csrs/index'
 import {OrganisationalUnit} from '../../../src/csrs/model/organisationalUnit'
 import {PageResults} from '../../../src/learning-catalogue/model/pageResults'
